Put all GLDebugger functionality inside GLProfiler
