+++ gdx/src/com/badlogic/gdx/graphics/profiling/GLProfiler.java
@@ -16,16 +16,24 @@
 
 package com.badlogic.gdx.graphics.profiling;
 
+import static com.badlogic.gdx.graphics.GL20.GL_INVALID_ENUM;
+import static com.badlogic.gdx.graphics.GL20.GL_INVALID_FRAMEBUFFER_OPERATION;
+import static com.badlogic.gdx.graphics.GL20.GL_INVALID_OPERATION;
+import static com.badlogic.gdx.graphics.GL20.GL_INVALID_VALUE;
+import static com.badlogic.gdx.graphics.GL20.GL_OUT_OF_MEMORY;
+
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.math.FloatCounter;
 
-/** This class will collect statistics about the GL calls. All calls to it will get counted and delegated to the actual GL20 or
- * GL30 instance.
+/** When enabled, collects statistics about GL calls and checks for GL errors.
+ * Enabling will wrap Gdx.gl* instances with delegate classes which provide described functionality
+ * and route GL calls to the actual GL instances.
  * 
  * @see GL20Profiler
  * @see GL30Profiler
  * 
- * @author Daniel Holderbaum */
+ * @author Daniel Holderbaum
+ * @author Jan Polák */
 public abstract class GLProfiler {
 
 	/** All calls to any GL function since the last reset. */
@@ -41,8 +49,30 @@ public abstract class GLProfiler {
 	public static int shaderSwitches;
 
 	/** The amount rendered vertices since the last reset. */
-	public static FloatCounter vertexCount = new FloatCounter(0);
+	public static final FloatCounter vertexCount = new FloatCounter(0);
 
+	public static String resolveErrorNumber (int error) {
+		switch (error) {
+		case GL_INVALID_VALUE:
+			return "GL_INVALID_VALUE";
+		case GL_INVALID_OPERATION:
+			return "GL_INVALID_OPERATION";
+		case GL_INVALID_FRAMEBUFFER_OPERATION:
+			return "GL_INVALID_FRAMEBUFFER_OPERATION";
+		case GL_INVALID_ENUM:
+			return "GL_INVALID_ENUM";
+		case GL_OUT_OF_MEMORY:
+			return "GL_OUT_OF_MEMORY";
+		default:
+			return "number " + error;
+		}
+	}
+
+	/** This listener will be called when GLProfiler is enabled and any GL call sets an error number (retrievable by glGetError call).
+	 *
+	 * Default is {@link GLErrorListener#LOGGING_LISTENER}. */
+	public static GLErrorListener listener = GLErrorListener.LOGGING_LISTENER;
+	
 	/** Enables profiling by replacing the {@code GL20} and {@code GL30} instances with profiling ones. */
 	public static void enable () {
 		if (!isEnabled()) {
@@ -64,7 +94,8 @@ public abstract class GLProfiler {
 		return Gdx.gl30 instanceof GL30Profiler || Gdx.gl20 instanceof GL20Profiler;
 	}
 
-	/** Will reset the statistical information which has been collected so far. This should be called after every frame. */
+	/** Will reset the statistical information which has been collected so far. This should be called after every frame.
+	 * Error listener is kept as it is. */
 	public static void reset () {
 		calls = 0;
 		textureBindings = 0;
