+++ tests/gdx-tests/src/com/badlogic/gdx/tests/utils/GdxTests.java
@@ -130,7 +130,7 @@ public class GdxTests {
 		ControllersTest.class,
 		Gdx2DTest.class,
 		GestureDetectorTest.class,
-		GLDebuggerTest.class,
+		GLProfilerErrorTest.class,
 		GroupCullingTest.class,
 		GroupFadeTest.class,
 		GroupTest.class,
